backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.6
alembic==1.14.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# AI / Embeddings
openai==1.58.1

# PDF processing
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5

# Configuration
pydantic-settings==2.7.1

# Email
email-validator==2.2.0